# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Loombre API
  version: 1.0.0
  summary: >-
    Contract-first API for Loombre, a ground-up self-hosted media streaming
    platform (movies, TV, music).
  description: >-
    This document is the source of truth (CLAUDE.md architecture invariant
    #1): apps/server controllers conform to it and packages/sdk is generated
    from it — never hand-authored.


    Conventions (apply globally; see components/ for shared shapes): all
    timestamps are BIGINT epoch milliseconds as JSON integers with field
    names suffixed `Ms`; all ids are UUIDv7 strings; all list endpoints use
    cursor pagination (`{ items, nextCursor }`, with `?cursor=&limit=` query
    params); all 4xx/5xx responses are RFC 9457 `application/problem+json`;
    enums are closed OpenAPI enums; core entity schemas set
    `additionalProperties: false`.


    Evolution policy: additive-only within a major version (docs/PLAN.md
    §4.1) — CI runs oasdiff against `main` and fails the build on any
    breaking change classification. Deprecated operations carry
    `deprecated: true` plus a `Sunset` response header for two minor releases
    minimum before a major bump.
  license:
    name: AGPL-3.0-relicense-pending
    identifier: LicenseRef-Loombre-Private
...
servers:
  - url: /v1
    description: Loombre API, version 1 (path-prefixed; host is deployment-specific)
security:
  - bearerAuth: []
...
paths:
...
  /admin/plugins/preview:
    post:
      operationId: previewAdminPlugin
      tags: [admin]
      summary: Fetch and validate a plugin's manifest without registering it (admin)
      description: >-
        C4's confirmation-screen data source: fetches GET <url>/lpp/manifest
        through the same SSRF-guarded transport registration itself uses,
        runs it through the staged manifest parser, and returns a summary —
        every declared capability with its scope, the full configSchema, and
        the union of requested event types. Nothing is persisted; calling
        this twice for the same URL never conflicts with anything, and
        nothing here mints a secret. An unknown capability type in the
        manifest fails with 422 whose `detail` states plainly that this
        Loombre does not support that capability type yet (C2); a
        SSRF-rejected URL (private/loopback/disallowed address, unless
        `lanAllowlist` covers it) also fails 422.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreviewPluginRequest' }
      responses:
        '200':
          description: Validated manifest summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PluginManifestPreview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins:
    get:
      operationId: listAdminPlugins
      tags: [admin]
      summary: List registered plugins, including health/breaker state (admin)
      description: >-
        Deliberately NOT cursor-paginated (mirrors listCrashFiles' documented
        deviation from the cursor-everywhere rule): a self-hosted instance's
        plugin registry is a small, admin-managed set, not an unbounded
        catalog collection.
      responses:
        '200':
          description: Registered plugins
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPluginList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      operationId: registerAdminPlugin
      tags: [admin]
      summary: Register a plugin (admin)
      description: >-
        LD6's registration state machine: re-fetches and re-validates the
        manifest at `url` (never trusts a client-supplied preview payload),
        validates the submitted grants against it, validates `config`
        against the manifest's configSchema (secret fields route to the
        keyring, never to the stored plugin row), mints the delivery-signing
        HMAC, runs a health check, and commits the plugin enabled with the
        granted scope. `hmacSecret` in the response is shown EXACTLY ONCE —
        it is never retrievable again through any endpoint. A failed health
        check does NOT block registration: the row still commits, with
        `healthState: unhealthy`; the admin UI is expected to surface this
        and offer enable-anyway vs cancel (the plugin already exists at that
        point — cancel means the UI immediately calling removeAdminPlugin).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterPluginRequest' }
      responses:
        '201':
          description: Plugin registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisterPluginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: A plugin is already registered at this baseUrl
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    get:
      operationId: getAdminPlugin
      tags: [admin]
      summary: Read one registered plugin (admin)
      responses:
        '200':
          description: The plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: removeAdminPlugin
      tags: [admin]
      summary: Remove a plugin (admin)
      description: >-
        LD9: deletes the plugin row (plugin_event_grants CASCADE) and every
        keyring entry it owns — the HMAC and every secret config field.
        Nothing keyring-side survives.
      responses:
        '204': { description: Plugin removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/config:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    put:
      operationId: updateAdminPluginConfig
      tags: [admin]
      summary: Update a plugin's config values (admin)
      description: >-
        Validated against the plugin's stored manifest configSchema exactly
        like registration; secret fields route to the keyring, never to the
        plugin row's non-secret `config`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePluginConfigRequest' }
      responses:
        '200':
          description: Updated plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/event-grants:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    put:
      operationId: updateAdminPluginEventGrants
      tags: [admin]
      summary: Update which requested event types this plugin is granted (admin)
      description: >-
        Full-replacement grant set for the plugin's already-GRANTED
        event-subscriber capability — every value must be a member of that
        capability's manifest-declared `eventTypes` request. Does not change
        which capability TYPES are granted (see reapproveAdminPlugin for
        that) and can never widen scope beyond what registration/reapproval
        already approved.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePluginEventGrantsRequest' }
      responses:
        '200':
          description: Updated plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This plugin's event-subscriber capability is not currently granted
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/pseudonymization:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    put:
      operationId: updateAdminPluginPseudonymization
      tags: [admin]
      summary: Toggle whether this plugin receives pseudonymous or real actor ids (admin)
      description: >-
        LPP v1 mission §3.2 (Lane W5b): plugins.pseudonymize_actor_ids
        (migrations/0016_plugin_delivery_cursors.sql), default TRUE — every
        user-id-bearing field the outbox delivery loop's actor-field map
        names for a delivered event is replaced with a per-(plugin,user)
        stable pseudonym before signing and delivery. Setting `enabled:
        false` sends real account ids to this plugin instead, effective
        immediately for every batch delivered from that point on (already-
        delivered batches are unaffected — this is not retroactive).
        Requires this plugin to currently hold the event-subscriber
        capability grant (409 otherwise, same guard as
        updateAdminPluginEventGrants) — the setting has nothing to act on
        for a plugin that never receives the activity feed.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePluginPseudonymizationRequest' }
      responses:
        '200':
          description: Updated plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This plugin's event-subscriber capability is not currently granted
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/enable:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      operationId: enableAdminPlugin
      tags: [admin]
      summary: Re-enable a manually- or breaker-disabled plugin (admin)
      description: >-
        Resets the health breaker. A plugin currently disabled for
        `disabledReason: scope-change` cannot be re-enabled this way —
        reapproveAdminPlugin is the only door out of that state (a plain
        re-enable would skip the new grant validation that state exists to
        force).
      responses:
        '200':
          description: Updated plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This plugin is disabled for a scope change and requires reapproveAdminPlugin
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      operationId: disableAdminPlugin
      tags: [admin]
      summary: Manually disable a plugin (admin)
      responses:
        '200':
          description: Updated plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/refresh:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      operationId: refreshAdminPlugin
      tags: [admin]
      summary: Re-fetch a plugin's manifest and apply/flag any scope change (admin)
      description: >-
        Any expansion (a new capability type, broader mediaKinds, a
        contentClass widened general -> restricted, a broader eventTypes
        request) auto-disables the plugin (`disabledReason: scope-change`)
        and leaves the previously-approved grants untouched pending
        reapproveAdminPlugin — `expanded: true` and `reasons` describe
        exactly what grew. A non-expanding refresh updates the stored
        manifest snapshot in place (grants only ever narrow automatically,
        never requiring re-approval).
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RefreshPluginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/reapprove:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      operationId: reapproveAdminPlugin
      tags: [admin]
      summary: Re-approve a plugin after a scope-expanding refresh (admin)
      description: >-
        Only valid while `disabledReason: scope-change`. Re-fetches the
        manifest fresh (time may have passed since the expansion was
        detected), validates the submitted grant against it exactly like
        registration, and re-enables the plugin with the new grant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReapprovePluginRequest' }
      responses:
        '200':
          description: Re-approved plugin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminPlugin' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: This plugin is not currently awaiting scope-change re-approval
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
  /admin/plugins/{id}/rotate-hmac:
    parameters:
      - $ref: '#/components/parameters/IdPathParam'
    post:
      operationId: rotateAdminPluginHmac
      tags: [admin]
      summary: Rotate a plugin's delivery-signing HMAC secret (admin)
      description: >-
        Mints a genuinely fresh secret, overwriting the stored one — the
        plugin must be reconfigured with the new value out-of-band. Returned
        EXACTLY ONCE, never retrievable again afterward.
      responses:
        '200':
          description: The new HMAC secret
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RotatePluginHmacResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT; refreshed via per-device rotating refresh tokens.
  parameters:
...
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or not visible under the caller's ViewerContext)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: Request failed schema/business-rule validation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Extension members are additive.
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable machine-readable error code, additional to `type`.
      required: [title, status]
      additionalProperties: true
...
    PluginManifestPreview:
      type: object
      description: >-
        previewAdminPlugin's result — a validated read of the plugin's
        manifest for the C4 confirmation screen. Nothing is persisted by
        that call.
      additionalProperties: false
      required: [name, version, protocolVersion, publisher, description, capabilities, configSchema, requestedEventTypes, manifestDigest]
      properties:
        name: { type: string }
        version: { type: string, description: The plugin's own version string — distinct from protocolVersion. }
        protocolVersion: { type: integer }
        publisher: { type: string }
        description: { type: string }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/PluginCapability' }
        # No `type: object` (mirrors AdminSettingSchemaEntry.default's bare
        # `{}` precedent): a real JSON-Schema document, opaque to this
        # contract, walked client-side by lib/plugin-manifest.ts.
        configSchema:
          description: >-
            The plugin's manifest configSchema verbatim (LPP §3's JSON-Schema
            subset) — the admin config form's sole input, auto-rendered by
            the same schema-driven widget renderer admin settings uses. A
            string field with `secret: true` is entered write-only and
            stored in the keyring, never echoed back.
        requestedEventTypes:
          type: array
          items: { type: string }
          description: Union of every event-subscriber capability's requested eventTypes (empty if the manifest declares none).
        manifestDigest:
          type: string
          description: >-
            C-2 fix wave (adversarial-review): a sha256 hex digest of the
            EXACT manifest content this preview validated. The client MUST
            round-trip this value back as `manifestDigest` on
            registerAdminPlugin / reapproveAdminPlugin — the server re-fetches
            the manifest at that point (as it always did) and 409s if the
            fresh fetch's digest no longer matches, closing the TOCTOU where
            a plugin could serve a broader/differently-scoped manifest to the
            registration call than the one this preview actually showed the
            admin.
    PreviewPluginRequest:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The plugin's base HTTP(S) origin — GET <url>/lpp/manifest is fetched.
        lanAllowlist:
          type: array
          items: { type: string }
          description: Explicit LAN hostnames/IPs to permit for this preview fetch even if they land in a private/loopback address range (LD5).
...
    AdminPluginList:
      type: object
      description: listAdminPlugins' result. Deliberately NOT cursor-paginated — see that operation's description.
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AdminPlugin' }
    RegisterPluginRequest:
      type: object
      additionalProperties: false
      required: [url, grantedCapabilityTypes, eventTypeGrants, config]
      properties:
        url: { type: string, format: uri }
        grantedCapabilityTypes:
          type: array
          items: { type: string }
          description: Subset of the manifest's declared capability `type` values to enable (LD6 "capability set <= declared").
        eventTypeGrants:
          type: array
          items: { type: string }
          description: Subset of the union of every event-subscriber capability's requested eventTypes to grant (LD6 "event grants <= requested").
        config:
          description: Raw submitted values keyed by configSchema property name — both secret and non-secret fields together (split server-side; secret values are never echoed back).
        lanAllowlist:
          type: array
          items: { type: string }
          description: Explicit LAN hostnames/IPs to permit for this plugin (LD5). Defaults to none.
        manifestDigest:
          type: string
          description: >-
            C-2 fix wave: the `manifestDigest` a prior previewAdminPlugin
            call returned for this exact plugin. Schema-optional (kept
            additive), but ENFORCED as required server-side — omitting it is
            a 422 ("preview this plugin first"), and a value that no longer
            matches a fresh re-fetch is a 409 (the manifest changed since it
            was previewed).
    RegisterPluginResponse:
      type: object
      description: >-
        hmacSecret is returned EXACTLY ONCE — it is never retrievable again
        through any endpoint (LD1). The admin UI must surface it immediately
        with a copy affordance and a "this will not be shown again" notice.
      additionalProperties: false
      required: [plugin, hmacSecret]
      properties:
        plugin: { $ref: '#/components/schemas/AdminPlugin' }
        hmacSecret: { type: string }
...